requirements-file.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pandas==2.0.3
streamlit==1.29.0
Pillow==10.0.0
dlib==19.24.2
